requirements.txt
django<5.3
environs
psycopg
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter

pytest
pytest-django